Added question for participants
Added a question for participants asking how many use or will be expected to use analytical software such as R, SAS, or Matlab? This question helps (I think) fill a hole in the introduction to this lesson - that hole being: why are these lessons so focused on getting data *out* of Excel?